fix: build benchmark against baseline libraries
