+++ benchmark/barback/cmd/bench/business_app_smoke_legacy.go
@@ -1,0 +1,14 @@
+//go:build benchmark_legacy
+
+package main
+
+import (
+	"context"
+	"fmt"
+
+	"go.mau.fi/whatsmeow"
+)
+
+func validateBusinessApp(context.Context, *whatsmeow.Client) error {
+	return fmt.Errorf("business app validation requires HyperMeow")
+}
